Arizona Credits Fixes #2460
