+++ policyengine_us/parameters/gov/states/az/tax/income/credits/amount.yaml
@@ -6,5 +6,5 @@ metadata:
   reference:
     - title: 43-1073. Family income tax credit (B)
       href: https://www.azleg.gov/ars/43/01073.htm
-value:
+values:
       2022-01-01: 40
