+++ policyengine_us/parameters/gov/states/az/tax/income/credits/eligibility/single.yaml
@@ -7,5 +7,5 @@ metadata:
   reference:
     - title: 43-1073. Family income tax credit (a) (9)
       href: https://www.azleg.gov/ars/43/01073.htm
-value:
-      2022-01-01: 10_000
+values:
+    2022-01-01: 10_000
