+++ policyengine_us/parameters/gov/states/az/tax/income/credits/max_amount.yaml
@@ -1,0 +1,20 @@
+description: Arizona provides the following maximum amount under the arizona family tax credit based on filing status.
+metadata:
+  period: year
+  unit: currency-USD
+  label: Arizona family tax credit maximum amount
+  breakdown:
+    - filing_status
+SINGLE:
+  2021-01-01: 120
+HEAD_OF_HOUSEHOLD:
+  2021-01-01: 240
+JOINT:
+  2021-01-01: 240
+WIDOW:
+  2021-01-01: 240
+SEPARATE:
+  2021-01-01: 120
+reference:
+  - title: 43-1073. Family income tax credit (B)
+    href: https://www.azleg.gov/ars/43/01073.htm
